git-svn-id: svn://123.57.173.36/meijia-parent@609 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/dao/user/UsersMapper.java
@@ -15,9 +15,9 @@ public interface UsersMapper {
 
 	int deleteByPrimaryKey(Long id);
 
-    int insert(Users record);
+    Long insert(Users record);
 
-    int insertSelective(Users record);
+    Long insertSelective(Users record);
 
     List<Users> selectByAll();
